+++ src/main/java/duke/DeleteException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DeleteException extends Exception {
     public DeleteException() {
         super("There is nothing to delete here!");
