+++ src/main/java/duke/Duke.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.*;
 import java.util.Scanner;
